+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Unreleased
+### Added
+- feat: Add `RTMCountryFlag` widget to facilitate the display of a flag
 
 ## 0.1.0 - 01/01/2025
 ### Added